python_backend/requirements.txt
web3
fastapi
uvicorn
python-dotenv
toml
py-solc-x
python-multipart
pylint
psycopg2-binary